SNAP-110: Added support to edit parameters in coinfiguration files that do not start in column 1.
Make use of the new feature to edit the "seeds" in the cassandra.yaml file.
Prevent uncommenting out a section name because we do not yet have the proper code to handle such a case.
Added flag so we can get the trimmed data reomved from the output file (i.e. comment introducer).
Added a parameter in the call in the cassandra.cpp file so we can remove spaces and dash to the left of the "seeds:" field.
Documented the function parameters in details.