SNAP-110: Re-enforce the SSL setup by creating an ssl-snap sub-directory and placing a default SSL configuration file in there.
Updated the default setup with our new paths/sub-paths and added the Include of snap-apache2-ssl.conf file.
Added a rule to make sure that the permissions of the ssl-snap directory gets set to 700 (only root can access/read/write)
Added zip and unzip and tools to be installed. It's so common that one needs that tool that it is not a bad thing to have it around.