SNAP-443: snapcommunicator uses PEM files by default now. One can remove the use by setting both ssl_certificate and ssl_private_key to empty.
Similarly the snap.cgi and snapserver can be setup to use PEM. The snapserver has an ssl_certificate and ssl_private_key set of variables and snap.cgi has a use_ssl variable.
snap.cgi and snapserver ignores the PEM files if the IP address of snapserver is 127.0.0.1.
Like snapcommunicator, it is possible to set ssl_certificate and ssl_private_key to empty so the server uses PLAIN instead of SECURE in that case.
Various changes so the rest of the projects compile with the new bio_server in place.