SNAP-443: Fixed the connect_cassandra(), if already connected, we must return true otherwise snapdbproxy does not do the right thing.
Fixed a bug in disconnect_cassandra(), we keep a pointer to the "sites" table and thus must reset it to really disconnect.
Replaced the create_table() with a get_table() and do not call a server create_table() function (i.e. it has to be created with the snapcreatetables tool ahead of time.)
Noticed that the snapbackend tool would not attempt to connect to Cassandra if it failed at the time the process received CASSANDRAREADY. Now it will try once per second until it works (again).
Various cleanups.