SNAP-440: Replaced perl with sed and now making a backup.
Since we have to parse the YAML in cassandra.yaml, the libraries we have
available strip all of the comments. So we will make sure the bundle
makes a backup (cassandra.yaml.bak) after the changes (and will preserve
the original file installed by the package with a .orig suffix).