SNAP-289: Added many try/catch blocks to various tests.
Moved the adgetopt initialization to the try/catch block whenever possible.
Added a logger_exists() function so we can do a try/catch around the log4cplus::Logger::exists() function since it can throw and we call it from a destructor. Plus, as far as we're concerned, if it throws it is the same as if the function returned false.
In snap_communicator, get the socket and make sure it is valid (>= 0) before using it.
Do a cast back to bool in the snap_expr function. At this point I just put a comment about the concerning "wrong use of bool".
Fixed a bug in the snap_communicator server_bio::accept() implementation. I was using the accept socket to set the SO_KEEPALIVE flag, instead of using the new client socket.
Removed the possibility of having a contrib sub-directory with catch in it.
Catch enhanced clara.h quite a bit, it means we need to use Catch::Clara::<classes> now and the bind() feature changed.
Fixed the new version of the libsnapwebsites library.
Changed the return value of the test_ssl/shutdown_client/server in case an exception occurs. It now exits with 1 when such an error occurs.
Removed the tabs in the ip_test.cpp file. Also added a try/catch there.
Added several tests again functions such as chmod() as pointed out by Coverity.