SNAP-440: Lots of improvements!
Fixed the bundle for cassandra so that it properly alters the
cassandra.yaml file. Added a postrm which deletes the entire
/etc/cassandra directory (since we create the ssl and public dirs)

Added support for yaml-cpp to read and write the cassandra.yaml file. It
does strip out all of the comments, but for that we make a copy in a
".orig" file so you can use the comments for reference.

Keys appear to be properly copied for snapdbproxy, for when it asks the
cassandra plugin for a public SSL key. I have yet to very that the
daemon actually works as expected, but this negotiation is at least
correct.