SNAP-440: Now keys are distributed from both ends.
The issue was that the public (or trusted) SSL keys for Cassandra were not
being properly distributed. Now, when snapdbproxy and the cassandra
manager plugin are restarted, they both send requests. Snapdbproxy sends
a CASSANDRAKEYS request, which causes the cassandra plugin to send
client keys, and the cassandra plugin sends CASSANDRASERVERKEY, so all
cassandra nodes on the network are informed of a potentially new key.

This will make sure everyone has the right keys for communication via
SSL.

Added the service name and server name to the trace statement in
snapcommunicator.