SNAP-440: turn off cassandra reboot after every change
When making changes to the cassandra YAML file, I turned off the
immediate restarting of the cassandra server. This is because you don't
want to restart with the new values until *all* of the addresses are set
(rpc_address/listen_address/broadcast). Also, you might have to make
sure encryption is turned on. So I added a "Restart Cassandra" button.
Also, I added a "Restart Everywhere" button so that way all of these
things could be changed at once.

Also, the encryption options allow a "save everywhere" for convenience.