SNAP-440: Added purge-keys ability for cassandra.
We now offer the option through snapmanagerdaemon to purge the SSL keys
for the cassandra server (also, there is a "Save Everywhere" that allows
one to do this all at once). Also, snapdbproxy now understands the
"force" parameter so that when the keys are shipped again through
snapcommunicator, the key will be overridden instead of ignored, should
a key from that server be already on the system.