SNAP-378: Changed the do_update() so it can only be called through the snapinstallwebsite--which actually works exactly the same for the first installation and updates.
The updating capability first sets the website state to "updating", then it waits 10 seconds assuming that is enough for all clients (i.e. cheap drainage), then it runs all the do_update() functions.
If the current state of a website is not "ready," frontend users get a 503 error. Thus we avoid potential conflicts with the do_update() process.
Added support for specifying the protocol on the snapinitializewebsite tool (i.e. --protocol HTTPS), so an administrator can have more than just 443 as a secure port.
Added my cassandra-start script to the snapdbproxy/tools, but that's not expected to be installed, this is just for programmers.
Removed access to the install_layout() call from frontend users. This would quickly auto-update layouts, but was very dangerous if run by more than one user.