SNAP-468: Fixed the new layout.cpp installation. I also needed to break that installation up in two parts. (install content.xml and then check the validity in the dynamic update)
Fixed the installation of the "default" website. With my previous changes, it got removed so it broke the installation of brand new websites (the only way to install the "default" theme was to have it done the first time it was requested and that code got removed.)
Fixed the snap_child::set_site_parameter() function. Since we removed the "create table" from the equation, we cannot "randomly" add a site name (it would overwrite the user's entered site name on each restart of the server.)
Moved that 'set website name' command to the initialization code and only apply it if the name is currently empty or undefined.
Added some log statements to better track installation behavior (since this has become a background command I'm not worried about speed.)
Changed the snap_communicator so errors of disconnected sockets are now either warnings or debug messages.
Added support for a query string in snapinstallwebsite so we can install a layout.
Added support for the install-layouts query string in the website installation code so we can request layouts to be installed.
Updated the snaplock daemon so when used with --list it does not show but errors and fatal errors.
Removed the layouts from under /admin/ and moved them to just /layouts/...
Removed the test to avoid changing the action to "administer" if it starts with /admin/layouts (SNAP-335).
Changed the date of the default theme to use the date in UTC (the __DATE__ and __TIME__ from the compiler are in local time.)
Added the extensions to the layout entries. This will make a lot more sense long term.
Fixed the layout last updated functions so they save a per layout timestamp instead of just one for the plugin + all layouts.
Fixed the ownership of a three pages to users_ui as expected by the users_ui::fix_owner_update() function.
Various clean ups.