SNAP-266: Added a new appender for log4cplus which directs to snapcommunicator.
The new appender allows the logging feature to not only log to file, but
also to redirect across the network to the yet-to-be-written "snaplogd"
service, which will write all received logs into one big file (or
perhaps multiple files depending on where it came from...?).

So far, snapserver and snapbackend have this new feature. I need to go
through and turn it on for those services we want to capture (basically,
all of them).