SNAP-355: Redesigned the iplock tool so we can actually use various schemes.
Changed the --block and --unblock so they are not cumulative anymore.
Added a --count command line option so we can now check whether a blocked IP is still in use.
Reorganized the project so we have sub-directories: conf, src where added.
Moved the version declaration in a version.h.in file.
Moved the poison to a separate file.
Bumped version to 3.x since it does not work exactly as before.