SNAP-355: We now properly broadcast BLOCK messages so all computer block those offensive IP addresses.
Cleaned up the snapfirewall implementation by adding a sub-class that handles the scheme, IP, and block period.
Renamed the "ip" parameter as "uri" because we now allow the inclusion of a scheme.
Added support for saving BLOCK data until we get a connection to the database.
Applied a fix to the next_wakeup() function. If no IP exists, we now sleep until the next BLOCK message.