SNAP-266: Added a custom snapcommunicator for the snap_child object.
The issue is that when the child forks, it no longer has access to the
g_communicator static object (that the server creates in an unnamed
namespace). This means that it does not have an active snapcommunicator
connection. We need to create a new messenger instance, and then pump
the queue to keep it active (by calling snap_communicator::run(), which
blocks).

This should allow a snap_child to forward logs through to snaplog.