SNAP-454: Fixed the destruction of clients so it happens as soon as we reset them in the thread runner by NOT keeping a copy in the main thread which has no use for it.
Fixed the passing of the BIO client shared pointer. At has to be references everywhere. Then the connection becomes the sole owner (using a swap() in the main thread.) That way the callers cannot do anything to it. This is important since we are dealing with a thread.
Created a close() function to clean up all those f_socket = -1 and reset() calls, because the f_socket = -1 actually needs to be guarded.
Various cleanups.