SNAP-461: Added a snaplock flag in the snapserver and snapbackend so they know whether the snaplock service is ready or not.
Added support for the STATUS message in the snapserver and snapbackend to get the status info of all services from the snapcommunicator.
Added a SERVICESTATUS message so one cna request for a STATUS message from the snapcommunicator for a specific service.
Added a send_status() call to the DISCONNECT message since that connection just went down.
Various cleanups.