SNAP-266: Feature complete, now to test.
I have mysql implemented into snaplock now. All logs should be added to
the database, and we will be able to query what we have through mysql.

In terms of root password, I've decided to not store it in the postinst
file (for security reasons). Now a password is generated using pwgen,
and stored in /var/lib/snapwebsites/private/snaplog_password.sh. The
"private" subdir is 0700 for root. This way, we keep other users out of
it.