SNAP-465: Adding fail2ban to check log files (iptables.log for now) and block IP addresses (useful to avoid the logs in this case.)
Added an logrotate file to rotate iptables.log files.
Added a fail2ban action to transform a jail in a BLOCk message to snapfirewall.
Added a fail2ban filter to check for IP addresses repetition in the iptables.log file.
Added a fail2ban jail to glue everything together.
Added a "Before=fail2ban" in case it wants to ban IPs with "snapfirewall".
Added a protection so the BLOCK message cannot stop the snapfireball daemon.