SNAP-493: Added error exception so we can determine when Cassandra disconnects.
The issue was that if the Cassandra server goes down, the driver does
not know how to automatically connect. When we recieve
CASS_ERROR_LIB_NO_HOSTS_AVAILABLE, that means that the remote server has
dropped the connection. However, it could be coming back up, so we need
to tell the proxy server to re-connect. I had to add a pointer to the
snapdbproxy object into the snapdbproxy_connection class. This allows me
to call snapdbproxy::no_cassandra() to force the timer to restart.