SNAP-465: Call the snapapacheinit script to run the common initialization.
Remove the site configurations on a remove as well.
Fixed a set of rmdir, we must make sure the directories exist.
Moved the snapapacheinit script to /usr/lib/snapwebsites/scripts/.
Moved the apache2 setup in snapapacheinit.
Merged all the apache2 initialization in snapapachelog.sh and then renamed the script snapapacheinit.