SNAP-110: Extracted user modified apache2 options to separate files.
Updated the concerned scripts accordingly.
Enable the rewrite module because we now use it for the redirect.
Replaced abbreviated flags with full length names in RewriteRule's.
Make use of the a2query to know whether a site is enabled.
Added the -m option to a2en* calls.
Various clean ups.