SNAP-110: Fixed the dbproxy plugin so we check the use_ssl flag as defined in snapdbproxy.conf.
Adding a debug message in snapmanagerdaemon plugins when connecting to cassandra to know whether we are using SSL or not.
Fixed the saving of field "cassandra_use_ssl", it has to return true and we do not have to send the RELOADCONFIG twice.
Fixed the mark_done() call in snapdbproxy to use "true" (i.e. kill the messenger as well once done.)
Various clean ups.