SNAP-465: Fixed the test to block proxy attempts received by snap.cgi.
Changed the block to 1 year on a proxy attempt.
Added a test, if snap.cgi receives /xmlrpc.php, we block that IP immediately.
Clearly marked that block_ip()'s default period is "day".
Added '\d+' to the failregex of the badrequest.
Added a new 'unwanted' apache jail for valid requests we do not want.
Actually install the new count-unwanted script.
Removed the 'sudo' from the count-unwanted and check that we are root.
Made 'bc' a dependency of snapfirewall since we use it in a script.
Changed the BLOCK behavior, accept an undefined period as "day".