SNAP-110: Put "true" in CRON_DAILY_RUN.
Set "true" in CRON_DB_UPDATE so we won't have to run --propupd
Set "true" in APT_AUTOGEN so it gets updated when upgrading/installing.
Set "19" in NICE so it is less likely to bother other processes.
Moved the REPORT_EMAIL to /etc/default/rkhunter.
SNAP-465: Include '[iptables] ' in the IPv6 rules.