SNAP-484: Prevent reinterpretation of the POST after a log in.
Added a snappassword.log file to log output of snappassword.
Added a corresponding logrotate entry.
exit(2) if password file cannot be read or user cannot be found.
Fixed the test of the exit code using proper macros.
Added a monthly deletion of session files.
Added a '\n' at the end of the session file.
Added a link back to home when outputting an error.