SNAP-509: Ignore all parameters on command line of snap.cgi and snapmanager.cgi.
Changed snapmanager-apache2.conf to use RewriteRule which prevents unwanted access (because all is sent to snapmanager.cgi which is in control).
SNAP-110: Moved snapmanager.cgi to /var/www/snapmanager/cgi-bin so it is not universally accessible.
Moved snap.cgi to /var/www/snap/cgi-bin so it is not universally accessible.
Added a default website for snap, although it should not be accessible.
Further broke up the Apache2 settings so it can be shared.