SNAP-506: Fixed error in callback.
This is the danger of reinterpret-casting a void* pointer to a known
type, but I have to work with the Cassandra C driver on this one. I made
a mistake of trying to convert a pointer of CassFuture* to future*. This
works correctly now and the callbacks work right now.