SNAP-515: iplock functions correctly (except for unblock).
Batch processing is working correctly, along with the other stand-alone
features (except for --unblock, which does not actually remove the rule
	  even though it looks correct).

Also, added the ipaddrs.txt.<pid> to go into a root-only accessible
folder in the "/etc/iplock" folder. I will likely change this to be in the
"/var/lib/iplock" root folder.