SNAP-515: Fixed design error to handle varying schemes in input file.
The input file for the --batch command is in the form "[IP] [scheme]".
However, what I had done the first time was to limit the --batch command
to forcing a --scheme keyword, and I ignored any line with the scheme
other than what was specified on the command line. The problem with that
is that each time you run the iplock tool with a different scheme, the
previous rules would be flushed.

Also, tying the --flush command to a scheme makes no sense since you can
only flush a table, and that's it. You cannot flush rules tied to an
interface or port. You flush all or nothing.

Essentially, this is my fault because I misunderstood the requirements
for SNAP-515, I reworked the logic. This means you can supply a file
full of ip addresses and schemes (http, all, smtp), and this newer way
can handle them all.