Fixed errors in snapbackup.
When attempting to dump a table that appears in the tables to ignore
list, the code that was supposed to make an exception for the switch
"--tables" didn't work. Now this functions properly.

Also, fixed a bug where when you are incrementally updating the backup,
the schema table would be appended over and over again with the same
tables each time. I changed it so that the "name" column is a primary key, and
took out the auto-generated "id" column.