SNAP-258: Now all accesses to the user table/row/cells are encapuslated.
You must use the `user_info_t` interface to access columns pertaining to
any and all user data. Also, the `user_key` which offers access to the
table itself, is now encapsulated as much as possible. So when you
access the user by email, the way the row is accessed "under the hood"
is hidden from the user. This sets us up for the next change: where we
change the row key from a string (user_key) to an id (int64_t).