SNAP-529: Properly clear the cache of the context when disconnecting.
Added a reconnect timer so if we lose our connection to Cassandra we can reconnect at some point instead of crashing.
Added a cassandra timer so the snapbackend without an action will exit automatically if it never receives the CASSANDRAREADY message.
Added a few try/catch so if a Cassandra error occurs, we can reset the connection and go on instead of crashing.
EX-158: Added a couple new functions to the thread FIFO.
Reduced the restart timeout for pagelist to 5min instead of 1h.
Various clean ups.