SNAP-258: I have converted user_key (i.e. email) to the int64 id.
I added a dynamic update which should convert all instances of the user
key to using an identifier instead. Each access of a user record is by
id now instead of by the canonical email address of the user. This opens
up an easier system so we can change email addresses.