SNAP-258: better architecture which drops the old rows and adds anew.
The old way was not working because of a fundamental misunderstanding I
was operating on. You cannot "insert" an existing row into a new id--the
QCassandra object model is deceptive in that regard. I had to first copy
the information (each cell), then drop the entire row and create a new
entry fresh. Same for the *index_row*.