SNAP-530: Added AJAX support for checking the users session status.
Fixed the hit=transparent feature with attachments, favicon, images.
Avoid displaying the "Unauthorized" error message on a transparent access.
Made the users hit variable part of the plugin so we can check it at any time even after the cookies were checked.
Added a hit=check which just returns the session current status.
Changed the users.js timeout handling to use an AJAX request instead of a reload (1) because if still valid we should not reload and (2) the reload meant loading many files (i.e. favicon.ico) instead of just checking the session which takes no time.
Added support for an AJAX check of the user session.
Allow the process_cookies() to generate output.
Fixed libsnapwebsites docs SOURCE, the headers were missing.
SNAP-181: Adding the remove-dups script mainly for reference because it should not be required (it does not get installed either.)
Various clean ups.