SNAP-541: the `snapdb.conf` configuration file is no longer "mandatory."
The code was fixed such that when you tried to read a parameter, it
would attempt to load a config file and call exit(1) with a fatal error
message if the file didn't exist.

Now the snap_config code allows you to test the existence of the
configuration file and avoid the forced process exit at the same time.
This way, I can test to see if the conf file exists, and if not, don't
try to access the configuration.

Also bumped the change logs. In addition, I added a sample conf file in
`/etc/snapwebsites/snapwebsites.d/snapdb.conf`.