SNAP-258: User may successfully change their email address.
First of all, the user_info_t object needed to be non-const, so the
calling could modify the currently logged in user.
Also, I was misusing the "users::original_email" field, treating that as
the non-canonical, where the "users::current_email" was the canonical
(user_key). I changed that so that both are non-canonical, but the
original email address is the one the user signed up with, which remains
unchanged.