SNAP-329: Fixed login redirect problem by saving the referrer using the user identifier instead of a plain field name for all.
Removed the snap_child parameter from the user_info_t constructor and changed the get_snap() implementation in that regard. It makes it lighter.
Introduced the IDENTIFIER_VALID and IDENTIFIER_ANONYMOUS instead of -1/0.
Fixed signals to makes use of references instead of a copy of user_info (it could end up making many copies otherwise!)
Fixed the user_info_t::get_value() to return a reference to the value.
Fixed the SNAP_NAME_USERS_HIT_CHECK name, the "HIT" was missing.
Added a status to the "?hit=check" test: "soft".
Fixed several check against the anonymous path using == anonymous instead of isEmpty(), which was wrong.
Changed the email canonicalization by having it done at the time you call the get_user_key() function.
Updated code to the user_info_t implementation for added security.
Changed the get_user_path() to be both: base and standard. But especially check whether the user is defined, if not use "user" as the path.
Various cleanups.