SNAP-468: Fixed the loading of the XSLT and XML files from the layout.
Fixed the name of the them used to load the boxes.
Removed some legacy support which is really not required anymore.
Also get the name of the theme when loading an editor form.
Moved the change-email to /change-email to avoid being under /admin.
Removed the '[form::source]' reference in the /change-email page.
Added the auto-save="no" attribute in the change-email form.