EX-79: Added support for functions without parameters in snap_expr.
Call toupper() for the toupper() function, instead of tolower().
Properly interpret the backslash in parsing a string in snap_parser.
Added a website_uri() function in the content plugin.
Changed the layout_script handling to use our snap_expr instead of QScript.
Fixed the snapexpr tool, we need to connect to port 4042 now.