Changed functionality to now use an object called "FieldDiv".
This encapsulates all of the data relevant to the modified div. This
also allows me to better move data around to events. In theory, this
should allow one to have multiple processes going at once--but for now,
the click event for the button is ignored if there is anything modified.

The code is cleaner, more compact and far easier to debug.