Various fixes so it compiles with g++ 6.x.
Added Qt5::Core to the list of link libraries in mimetype.
EX-79: Changed the filter to pass the XML document in token_infot_t.
emoved the `javascript` plugin as a dependency of `layout` and `output`.
Put the layout-name and theme-name in the metadata instead of body.
Cleaned up the filtering process in the create_body() function.
Added a create_body_string() which will create the body and keep the CSS and JS links. This is for the `[content::page(...)]` token.
Various clean ups.