SNAP-185: Replaced the `jQuery()` accesses to "canonical" with "page-uri".
Added the `<link rel="page-uri" ...>` tag to the header.
EX-79: Fixed the box-sizing of the editor toolbar, it still appears at the wrong position, though.
Added the `expr()` function to the filter functionality.
Fixed the filter `token_help()` function with missing entries.
Fixed an error message in layout.cpp so it shows the correct filename.