EX-131: Added batch support to snapdbproxy/QCassandraOrder.
Batch support is already in libcasswrapper--so it was only a matter of
implementing it into libQtCassandra and snapdbproxy.

I added three new order statuses: BATCH_DECLARE, BATCH_ADD,
BATCH_COMMIT. This works with the existing libcasswrapper API (which
wraps the Cassandra CPP driver API). You declare the batch first (using
QtCassandra::QCassandraTable::startBatch()), then insert or delete as normal.
You must called the endBatch() method in order to send all of your edits
to the remote server.

This is untested code.