Moved libQtCassandra into libsnapwebsites.
This is done because there is no reason to keep this library under
`contrib`. The internal management has changed such that it relies on a
snap component to even work: snapdbproxy. It effect, it is intricately
tied to the snap library, and vice-versa. So now, the libsnapwebsites
package provides this library. At some point, I will rename it as
`libdbproxy`, which makes far more sense. It is no longer specific to
Cassandra.