EX-181: Added support for journaling new pages.
This design allows you to grab a new journal_list pointer (which is on
the top of the stack) and add page urls. Then you call its "done()"
method. This pops the list pointer off of the stack, but keeps the list
itself in a vector member in the content plugin. Once the final list has
its `done()` method called, the last list is popped and the urls are
processed.

As you add an url, immediately a new entry is created in the `journal`
table, along with a timestamp. If everything works right, content will
be created and those entries should be removed upon completion.

If anything fails to complete (like an exception is thrown during page
creation, for example), the vector of lists is not processed, and the
entries remain in the journal. The backend process examines the contents
of this list--if any entries are more than 5 minutes old, the url
contained in the entry is destroyed (using content::destroy_page()).