EX-181: Added the test itself to exercise the journal table.
The backend journal process method now accepts a parameter specifying
the number of minutes to wait until clearing the stale content and journal
table entries.

My test uses the backend processing, but does not rely on the backend
running. It might be a good idea to temporarily turn off your backend
cron task, as it could potentially conflict.

The test runs a happy path first, then does a second pass, without
calling a final `journal_test::done()` time. Then the backend is run,
and the journal table should be empty after that.