Added glob_dir class to snapwebsites lib, and use it in snapmanagercgi.
Added a glob_dir class to wrap the enumeration of the glob_t structure.
This is a convenience class so you don't have to remember to free the
C-based resources correctly each time. Just use the class, enumerate the
contenting using a C++11 lambda function, and it's smooth sailing.

Just remember to catch the exception generated from `glob_dir::set_path()`
on error.