Fixed XSLT translation in the layout with regard to token replacement.
Once the data is run through the filter, we also need to do one extra
run to find any tokens in the XSLT (e.g. '[plugin::replacement]'). If
they exist, we need to replace them before sending them to the XSLT
parser in Qt5--otherwise, the evaluation fails with an exception and the
page is blank.