EX-185 and EX-193: Make sure spaces aren't allowed, but allow comments.
To solve EX-185, I brute-forced ignored any spaces at all in the original email address.
This was a hack! I now see that the correct way is to wait until the parsing is
done and evaluate the parsed out username and domain name, and make sure they
have no whitespace in them. This solves both issues at once.