SNAP-597: rename QtCassandra to libdbproxy.
The name change is not only a namespace change, but a class name change.
Instead of having an unwieldy and long name, like
QtCassandraRowKeyPredicate, and just renaming it to be
libdbproxy_row_key_predicate, I decided to take the redundant part out
of the class names, since that's captured in the namespace (libdbproxy).