SNAP-597: Refactored schema, and added query name subbing.
Lots of changes, particularly in the schema classes--they are no longer
class-within-classes, but stand independently. They no longer point back
to their parents with a weak pointer. They have the code built into the
constructors to self-construct based on a meta class provided by the
cassandra-cpp-driver API. This is a cleaner way to construct, and the
code is much easier to understand.

Also, the QSql driver is able to do basic queries (but all with blocking
turned on, I'm afraid. I could not get non-blocking to work right, due
to the fact that the event that returns after the object is destroyed
tries to access a stale pointer, causing a crash).

Also improved the test structure somewhat.