SNAP-597: Improve and fix the issues with the QSql driver for Qt5.
Moved the driver code to a sister dir in the src folder (src/
qtcassandra).
Added recursive mutexes to prevent deadlocking in the same thread.
About to remove the query_model completely.
Driver now USEs the database, because that is how the Qt SQL models
query on tables.
Changed json names to mixed case.