SNAP-597: Improved the threading, and changed rowset to one PAGE.
I was over-broad in my locking, and wound up creating deadlocks. So I
made the locking as atomic as possible. There is still a deadlock
problem when the Qt5 sqlbrowser example project is used to load our new
driver when in Async mode. But running it in blocking mode seems to work
well.

However, I had it so the driver sucked up all of the records, page after
page. I now limit the initial exec of the query to one page worth (and
the QCassandraResult allows tuning the page size, which defaults to
100). Calling QSqlQuery::exec() with no arguments grabs the next page
worth. It returns false when at the end.