SNAP-597: Fixed the issue with fetching.
The model will "ping" to see if we can fetch up to 255 records.
So we do it, and then return as if we are on the 255th record. This
makes the model/view work smoothly with no locking up of the UI,
even on a remote connection with almost 100,000 records.