Added the --build and --no-build command line options to the build script.
Many changes to the libexcept project so the packages are cleaner and no references to libQtCassandra are present in the files.
Bumped version of the projects to recompile against the new version of libexcept.
Added a little bit of code in the SnapBuildConfig.cmake so we get a graph of the project dependencies (much easier to see where we're at.)
Added an enumeration function to the links plugin so we can go through all the children of a page without having to rewrite that code over and over again.