Updated script to generate output in one folder.
Now the origanization is:
   BUILD/Debug
   BUILD/Release
   BUILD/Sanitize

I still need to change all the ./mk script to use those folders. See the
one in snapbuilder as a version that will work automatically for 99% of
our projects.

This also means I removed the --sanitize flag since we can always create
that version (cmake is fast, it doesn't get compiled by default).