Started work on a tool to run builds on launchpad.
The snapbuilder tool reads the deps.make file which (1) gives it the list
of projects and (2) gives it the order in which to build the projects on
launchpad. The tool has a GUI with the list of projects and by rebuilding
one and marking it as having "a broken change", you can automatically have
all the other projects (those that depend on it) rebuilt.