Started implementing many of the features.
Added more info about launchpad build service.
Fixed the mk script, the LESS and TEST were reset after the while loop.
Fixed the test to know whether a push is required.
Fixed the sort() function, the project now appear in the expected order.
Fixed the project assignment operator, the & was missing.
Added a project_ptr so I can save the project pointers in the table cells.
Added a couple of widget to display the current selection name on the right.
Added messages in the statusbar so we know what's going on.
Fixed the selection on a per row basis and a single row at a time.
Properly reset the projects when refreshing the list from scratch.
Keep the currently selected row selected between refreshes.
Added a menu entry to allow for building the sanitized version.
Auto re-select item after a refresh.