Added a QSvgWidget to show the dependencies.
The widget is not showing the image ratio properly, so we have some work there...
Moved the Refresh buttons (column in table) as one button on the side.
Added a Coverage button (Remains disabled at the moment, we need to have a clean implementation of that one).
Very early implementation of the background processing class.